mount_linux: implement hanwen's poll hack for fixing test deadlock
The idea for this fix, and most of the code from it, should be
credited to hanwen, from commit 4f10e248e of the hanwen/go-fuse repo.

Briefly, on Linux, when the same process that's serving a fuse
mountpoint is sending epoll requests to that mountpoint (as is common
in tests), the go runtime can deadlock.  This works around the issue
with a nasty hack that forces the kernel to learn (and cache) the fact
that polling is not implemented right after the mount happens, to
avoid deadlocks when the test later makes the same request.

Issue: golang/go#21014
Issue: hanwen/go-fuse#165
Issue: bazil/fuse#183